requirements.txt
pygame
pillow
numpy
psutil
